renamed all components files to have .jsx extension
