+++ frontend/src/components/AppSearch/InstalledAppSearch.jsx
