+++ frontend/src/components/Apps/Figma/about-secttion/FigmaAbout.jsx
