+++ frontend/src/components/Apps/Figma/about-secttion/MainContent.jsx
