+++ frontend/src/components/Apps/Figma/components/FigmaTextBar.jsx
