+++ frontend/src/components/MainPage/ToolsHero1.jsx
