+++ frontend/src/components/Skeleton/components/CategoriesSection.jsx
