+++ frontend/src/components/Skeleton/components/DailyTools.jsx
