+++ frontend/src/components/Skeleton/components/EnterpriseTools.jsx
