+++ frontend/src/components/Skeleton/components/HeroSection.jsx
