+++ frontend/src/components/Skeleton/components/InstalledTools.jsx
