+++ frontend/src/components/Skeleton/components/SearchBar.jsx
