+++ frontend/src/components/Skeleton/components/TitleBox.jsx
