+++ frontend/src/components/Skeleton/components/Tool.jsx
