+++ frontend/src/components/Skeleton/components/WorkFromHome.jsx
