+++ frontend/src/components/Skeleton/components/installedcard.jsx
