+++ frontend/src/components/Skeleton/pages/ToolsView.jsx
