+++ frontend/src/components/searchbar.component.jsx
